delete first test client and rename second
